+++ clean_python/api_client/__init__.py
@@ -2,3 +2,4 @@ from .api_gateway import *  # NOQA
 from .api_provider import *  # NOQA
 from .exceptions import *  # NOQA
 from .files import *  # NOQA
+from .sync_api_provider import *  # NOQA